# refactor(Processor): remove unnecessary json.Marshalling while making common metadata from singular event