## Description of the change

> Removing `json.Marshal` from `makeCommonMetadataFromSingularEvent`.
> Left the function for now to facilitate a benchmark test.

with Marshal:
```Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^Benchmark_makeCommonMetadataFromSingularEvent_jsonMarshalling$ github.com/rudderlabs/rudder-server/processor

DB_DSN: postgres://rudder:password@localhost:55021/jobsdb?sslmode=disable
goos: darwin
goarch: arm64
pkg: github.com/rudderlabs/rudder-server/processor
Benchmark_makeCommonMetadataFromSingularEvent_jsonMarshalling-8   	  105403	     10408 ns/op	    4960 B/op	      70 allocs/op
PASS
ok  	github.com/rudderlabs/rudder-server/processor	9.073s
```
without:
```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^Benchmark_makeCommonMetadataFromSingularEvent_withoutMarshalling$ github.com/rudderlabs/rudder-server/processor

DB_DSN: postgres://rudder:password@localhost:55022/jobsdb?sslmode=disable
goos: darwin
goarch: arm64
pkg: github.com/rudderlabs/rudder-server/processor
Benchmark_makeCommonMetadataFromSingularEvent_withoutMarshalling-8   	 1519579	       759.0 ns/op	     544 B/op	       3 allocs/op
PASS
ok  	github.com/rudderlabs/rudder-server/processor	6.907s
```

## Notion Link

> [notion](https://www.notion.so/rudderstacks/Remove-marshalling-from-makeCommonMetadataFromSingularEvent-a1f78573d37b46579b820a1d05f7c30f)

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

## Checklist:

- [ ] My code follows the style guidelines of this project
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added unit tests for the code
- [ ] I have made corresponding changes to the documentation

## Security

- [x] The code changed/added as part of this pull request won't create any security issues with how the software is being used.